add rng_drand48 from perl5
and FreeBSD. Seperated into 64bit and 32bit implementations.
The double part is 2x slower than needed.
perl5 only uses the double part, which is weird given that they produce
only 48 bit not the needed 53 bit. They also have no special int method.
The seed is also not used to advance the state, not even on init.